various: unify control
