+++ crates/components/src/fullscreen/desktop.rs
@@ -1,7 +1,6 @@
-use super::controls::{ProgressBarControl, VolumeControl};
 use super::metadata::TrackMetadata;
 use super::tabs::Tabs;
-use crate::player_controls::{TransportButtons, TransportVariant};
+use crate::player_controls::{ControlsVariant, SeekSlider, TransportButtons, VolumeSlider};
 use crate::titlebar::Titlebar;
 use config::AppConfig;
 use dioxus::prelude::*;
@@ -74,14 +73,15 @@ pub(crate) fn FullscreenDesktop(
                     }
 
                     if !use_player_bar {
-                        ProgressBarControl {
+                        SeekSlider {
                             current_song_duration,
                             current_song_progress,
+                            variant: ControlsVariant::Fullscreen,
                         }
 
-                        TransportButtons { is_playing, variant: TransportVariant::Fullscreen }
+                        TransportButtons { is_playing, variant: ControlsVariant::Fullscreen }
 
-                        VolumeControl { player, config, volume, persisted_volume }
+                        VolumeSlider { player, config, volume, persisted_volume, variant: ControlsVariant::Fullscreen }
                     }
                 }
