+++ crates/components/src/player_controls.rs
@@ -1,3 +1,4 @@
+use crate::shared::fmt_time;
 use config::AppConfig;
 use dioxus::prelude::*;
 use hooks::use_player_controller::{LoopMode, PlayerController};
@@ -150,7 +151,7 @@ pub fn use_volume_mute(
 }
 
 #[derive(Clone, Copy, PartialEq)]
-pub enum TransportVariant {
+pub enum ControlsVariant {
     Fullscreen,
     Bar,
 }
@@ -167,9 +168,9 @@ struct TransportClasses {
     badge: &'static str,
 }
 
-fn transport_classes(variant: TransportVariant) -> TransportClasses {
+fn transport_classes(variant: ControlsVariant) -> TransportClasses {
     match variant {
-        TransportVariant::Fullscreen => TransportClasses {
+        ControlsVariant::Fullscreen => TransportClasses {
             wrapper: "flex items-center justify-between w-full mb-3",
             side: "w-11 h-11 rounded-full flex items-center justify-center transition-colors active:scale-95 relative flex-shrink-0 hover:bg-white/10",
             side_idle: "color: rgba(255,255,255,0.6);",
@@ -180,7 +181,7 @@ fn transport_classes(variant: TransportVariant) -> TransportClasses {
             play_icon_size: "text-4xl",
             badge: "absolute bottom-1 left-1/2 -translate-x-1/2 text-[10px] font-bold leading-none",
         },
-        TransportVariant::Bar => TransportClasses {
+        ControlsVariant::Bar => TransportClasses {
             wrapper: "flex items-center gap-2",
             side: "w-9 h-9 rounded-full flex items-center justify-center text-slate-400 hover:text-white hover:bg-white/10 transition-colors active:scale-95 relative flex-shrink-0",
             side_idle: "",
@@ -195,18 +196,18 @@ fn transport_classes(variant: TransportVariant) -> TransportClasses {
 }
 
 #[component]
-pub fn TransportButtons(is_playing: Signal<bool>, variant: TransportVariant) -> Element {
+pub fn TransportButtons(is_playing: Signal<bool>, variant: ControlsVariant) -> Element {
     let mut ctrl = use_context::<PlayerController>();
     let classes = transport_classes(variant);
     let inner_gap = match variant {
-        TransportVariant::Fullscreen => "flex items-center gap-4",
-        TransportVariant::Bar => "contents",
+        ControlsVariant::Fullscreen => "flex items-center gap-4",
+        ControlsVariant::Bar => "contents",
     };
 
     rsx! {
         div {
             class: classes.wrapper,
-            style: if variant == TransportVariant::Fullscreen { "max-width: 640px;" } else { "" },
+            style: if variant == ControlsVariant::Fullscreen { "max-width: 640px;" } else { "" },
             button {
                 class: classes.side,
                 style: if *ctrl.shuffle.read() { "color: var(--color-indigo-500);" } else { classes.side_idle },
@@ -250,5 +251,175 @@ pub fn TransportButtons(is_playing: Signal<bool>, variant: TransportVariant) -> 
                 }
             }
         }
+    }
+}
+
+#[component]
+pub fn SeekSlider(
+    current_song_duration: Signal<u64>,
+    current_song_progress: Signal<u64>,
+    variant: ControlsVariant,
+) -> Element {
+    let seek = use_seek_drag(current_song_duration, current_song_progress);
+    let display_progress = seek.display_progress;
+    let progress_percent = seek.progress_percent;
+    let is_radio = seek.is_radio;
+    let on_commit = seek.on_commit;
+    let on_input = seek.on_input;
+
+    match variant {
+        ControlsVariant::Fullscreen => rsx! {
+            div {
+                class: "w-full mb-3",
+                style: "max-width: 640px;",
+                div {
+                    class: "flex items-center gap-3",
+                    span { class: "text-xs text-white/70 font-mono", style: "width: 50px; text-align: left;", "{fmt_time(display_progress)}" }
+                    div {
+                        class: format!("flex-1 {} relative group", if is_radio { "" } else { "cursor-pointer" }),
+                        style: "height: 20px;",
+                        div {
+                            class: "absolute bg-white/20 rounded-full",
+                            style: "height: 4px; top: 8px; left: 0; right: 0;"
+                        }
+                        div {
+                            class: "absolute rounded-full pointer-events-none bg-white/90",
+                            style: "height: 4px; top: 8px; left: 0; width: {progress_percent}%;"
+                        }
+                        div {
+                            class: if cfg!(target_os = "android") {
+                                "absolute bg-white rounded-full pointer-events-none"
+                            } else {
+                                "absolute bg-white rounded-full pointer-events-none opacity-0 group-hover:opacity-100 transition-opacity"
+                            },
+                            style: "width: 12px; height: 12px; top: 4px; left: calc({progress_percent}% - 6px);"
+                        }
+                        input {
+                            r#type: "range",
+                            min: "0",
+                            max: "{*current_song_duration.read()}",
+                            value: "{display_progress}",
+                            class: format!("slider-hit absolute top-0 left-0 w-full h-full opacity-0 {}", if is_radio { "" } else { "cursor-pointer" }),
+                            disabled: is_radio,
+                            onchange: move |evt| on_commit.call(evt),
+                            oninput: move |evt| on_input.call(evt),
+                        }
+                    }
+                    span { class: "text-xs text-white/70 font-mono", style: "width: 50px; text-align: right;", "{fmt_time(*current_song_duration.read())}" }
+                }
+            }
+        },
+        ControlsVariant::Bar => rsx! {
+            div {
+                class: "flex items-center gap-2 w-full",
+                span { class: "text-[10px] text-slate-500 w-8 text-right font-mono", "{fmt_time(display_progress)}" }
+                div {
+                    class: format!("flex-1 h-1 bg-white/10 rounded-full relative {}", if is_radio { "" } else { "group cursor-pointer" }),
+                    div {
+                        class: "absolute top-0 left-0 h-full bg-white/90 rounded-full pointer-events-none",
+                        style: "width: {progress_percent}%",
+                        div { class: "absolute -right-1.5 -top-1 w-3 h-3 bg-white rounded-full opacity-0 group-hover:opacity-100 transition-opacity" }
+                    }
+                    input {
+                        r#type: "range",
+                        min: "0",
+                        max: "{*current_song_duration.read()}",
+                        value: "{display_progress}",
+                        class: format!("slider-hit absolute top-0 left-0 w-full h-full opacity-0 z-10 {}", if is_radio { "pointer-events-none" } else { "cursor-pointer" }),
+                        disabled: is_radio,
+                        onchange: move |evt| on_commit.call(evt),
+                        oninput: move |evt| on_input.call(evt),
+                    }
+                }
+                span { class: "text-[10px] text-slate-500 w-8 font-mono", "{fmt_time(*current_song_duration.read())}" }
+            }
+        },
+    }
+}
+
+#[component]
+pub fn VolumeSlider(
+    player: Signal<Player>,
+    config: Signal<AppConfig>,
+    volume: Signal<f32>,
+    persisted_volume: Signal<f32>,
+    variant: ControlsVariant,
+) -> Element {
+    let vol = use_volume_mute(player, config, volume, persisted_volume);
+    let volume_percent = vol.volume_percent;
+    let is_muted = vol.is_muted;
+    let toggle_mute = vol.toggle_mute;
+    let on_wheel = vol.on_wheel;
+    let on_commit = vol.on_commit;
+    let on_input = vol.on_input;
+
+    match variant {
+        ControlsVariant::Fullscreen => rsx! {
+            div {
+                class: "flex items-center gap-5 w-full",
+                style: "max-width: 640px;",
+                i { class: "fa-solid fa-volume-low text-white/40" }
+                div {
+                    class: "flex-1 cursor-pointer relative group",
+                    style: "height: 20px;",
+                    onwheel: move |evt| on_wheel.call(evt),
+                    div {
+                        class: "absolute bg-white/20 rounded-full",
+                        style: "height: 4px; top: 8px; left: 0; right: 0;"
+                    }
+                    div {
+                        class: "absolute bg-white/90 rounded-full pointer-events-none",
+                        style: "height: 4px; top: 8px; left: 0; width: {volume_percent}%;"
+                    }
+                    div {
+                        class: if cfg!(target_os = "android") {
+                            "absolute bg-white rounded-full pointer-events-none"
+                        } else {
+                            "absolute bg-white rounded-full pointer-events-none opacity-0 group-hover:opacity-100 transition-opacity"
+                        },
+                        style: "width: 12px; height: 12px; top: 4px; left: calc({volume_percent}% - 6px);"
+                    }
+                    input {
+                        r#type: "range",
+                        min: "0",
+                        max: "1",
+                        step: "0.01",
+                        value: "{*volume.read()}",
+                        class: "slider-hit absolute top-0 left-0 w-full h-full opacity-0 cursor-pointer",
+                        onchange: move |evt| on_commit.call(evt),
+                        oninput: move |evt| on_input.call(evt),
+                    }
+                }
+            }
+        },
+        ControlsVariant::Bar => rsx! {
+            div {
+                class: "flex items-center gap-2 group",
+                button {
+                    class: "w-9 h-9 rounded-full flex items-center justify-center text-slate-400 hover:text-white hover:bg-white/10 transition-colors active:scale-95",
+                    onclick: move |_| toggle_mute.call(()),
+                    i { class: if is_muted { "fa-solid fa-volume-xmark text-xs" } else { "fa-solid fa-volume-high text-xs" } }
+                }
+                div {
+                    class: "w-24 h-1 bg-white/10 rounded-full group/vol cursor-pointer relative",
+                    onwheel: move |evt| on_wheel.call(evt),
+                    div {
+                        class: "absolute top-0 left-0 h-full bg-white/90 rounded-full pointer-events-none",
+                        style: "width: {volume_percent}%",
+                        div { class: "absolute -right-1.5 -top-1 w-3 h-3 bg-white rounded-full opacity-0 group-hover/vol:opacity-100 transition-opacity" }
+                    }
+                    input {
+                        r#type: "range",
+                        min: "0",
+                        max: "1",
+                        step: "0.01",
+                        value: "{*volume.read()}",
+                        class: "slider-hit absolute top-0 left-0 w-full h-full opacity-0 cursor-pointer z-10",
+                        onchange: move |evt| on_commit.call(evt),
+                        oninput: move |evt| on_input.call(evt),
+                    }
+                }
+            }
+        },
     }
 }
